requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
supervision>=0.20.0
pydantic>=2.0.0
requests>=2.31.0
huggingface-hub>=0.20.0
accelerate>=0.27.0
bitsandbytes>=0.41.0
groundingdino-py>=0.4.0
